+++ Changes.md
@@ -1,4 +1,6 @@
 # master
+
+# 3.12.0
 - Emit doc commments `/** this is a doc comment */` in the TypeScript output.
 - Preserve the parameter names in function types emitted by the TypeScript back-end.
 - Treat a function with a single argument of unit type as if it had no arguments.